feat: [PIE-4272]: resolved circular dependancy
